+++ src/run.py
@@ -1,23 +1,5 @@
-import lo2yw7lib
+import lo2yw7
 
 
 def export_yw():
-    lo2yw7lib.export_yw()
-
-
-def to_blank_lines():
-    lo2yw7lib.to_blank_lines()
-
-
-def indent_paragraphs():
-    lo2yw7lib.indent_paragraphs()
-
-
-def replace_bullets():
-    lo2yw7lib.replace_bullets()
-
-
-def is_anything_selected(oDoc):
-    return lo2yw7lib.is_anything_selected(oDoc)
-    
-to_blank_lines()
+    lo2yw7.export_yw()